+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # PyBigBuy Changelog
 
-Note: the BigBuy API is not versioned and some endpoints are added or removed without notice. The removal of methods
-that call deleted endpoints is not considered a breaking change.
+Note: the BigBuy API is not versioned and some endpoints are often added or removed without notice.
+The removal of methods that call deleted endpoints is not considered a breaking change.
+
+## Unreleased
+
+* Add `get_user_auth_status`, `get_product_stock_by_handling_days`, `get_product_variations_stock_by_handling_days`
 
 ## 3.19.1 (2023/09/29)